POC: added Page component to provide basic UI for the application  - has a title row, content body, and footer section  - width and minimum height can be configured  - on_input callback is supported but untested
Page component is now applied to all views except the start screen.
Width is set to 90 chars, height to 32 chars.
When running from the built in terminal in PyCharm, you may want to resize the terminal vertically.
When running from the native terminal of the OS, the window should resize automatically (only tested on Windows)